Added new source files src/ptg.cpp and src/ptg.h
The new source files will hold the code for the path planning
 generation.  This will keep main.cpp a bit cleaner and things
will be easier to read for the graders.

src/ptg.cpp conatisn currently a lot of junk code which is out-
commented.  It is Python code taken from the lextured quizzes.

I also added the quintic polynomial solver JMT from the lecture
quizzes.

CMakeLists.txt now references the new source file such that make
compiles it automatically with the rest.